requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.7.1
Werkzeug==2.3.6
